Get a Position Frequency Matrices (PFM) with annotations

Using this function you can retrieve a Position Frequency Matrices (PFM) associated with a matrix profile Identifier along with its details and annotations. If a base ID (i.e. without version suffix) was supplied, the latest version will be returned.

Usage

rba_jaspar_matrix(matrix_id, file_format = NULL, save_to = NULL, ...)

Arguments

matrix_id

Character: A matrix profile Identifier. It has "base_id.version" naming schema.

file_format

Character: Instead of returning a R object, you can directly download the profile matrix in file with this format. Supported formats are: "yaml", "jaspar", "transfac", "meme" and "pfm"

save_to

NULL or Character:

NULL: (only if file_format was supplied) Save the file to an automatically-generated path.

Character string: A valid file or directory path to save the file to.

...

rbioapi option(s). See rba_options's arguments manual for more information on available options.

Value

A list that contains the PFM along with its details and annotations. If file_format was supplied, an un-parsed character string with the file's content.

Corresponding API Resources

"GET https://jaspar.elixir.no/api/v1/matrix/{matrix_id}/"

References

Rauluseviciute I, Riudavets-Puig R, Blanc-Mathieu R, Castro-Mondragon JA, Ferenc K, Kumar V, Lemma RB, Lucas J, Chèneby J, Baranasic D, Khan A, Fornes O, Gundersen S, Johansen M, Hovig E, Lenhard B, Sandelin A, Wasserman WW, Parcy F, Mathelier A JASPAR 2024: 20th anniversary of the open-access database of transcription factor binding profiles Nucleic Acids Res. in_press; doi: 10.1093/nar/gkad1059

Khan, A. and Mathelier, A. JASPAR RESTful API: accessing JASPAR data from any programming language. Bioinformatics, 2017, doi: 10.1093/bioinformatics/btx804

JASPAR API Documentation

Citations note on JASPAR website

See also

Examples

# \donttest{
rba_jaspar_matrix("MA0600.2")
# }
if (FALSE) { # \dontrun{
rba_jaspar_matrix(matrix_id = "MA0600.2",
                  file_format = "meme",
                  save_to = "my_matrix.meme")
} # }