Search matrix profiles available in JASPAR

You can use this function to list the JASPAR matrix profiles that match your search query, or run the function without any arguments to return a list of every matrix profile available in the latest release.

Usage

rba_jaspar_matrix_search(
  term = NULL,
  tf_name = NULL,
  tf_class = NULL,
  tf_family = NULL,
  tax_group = NULL,
  tax_id = NULL,
  data_type = NULL,
  collection = NULL,
  release = 2022,
  only_last_version = FALSE,
  order = NULL,
  page_size = 1000,
  page = 1,
  ...
)

Arguments

term

Character: A search term.

tf_name

Character: Transcription factor names (Case-sensitive).

tf_class

Character: Transcription factor class

tf_family

Character: Transcription factor family

tax_group

Character: Taxonomic group. Use rba_jaspar_taxons to get a list of supported Taxonomic groups.

tax_id

Numeric: NCBI taxonomic Identifier of species. Use rba_jaspar_species to get a list of supported Species.

data_type

Character: Type of the data (i.e The Methodology used for matrix construction). For example: "ChIP-seq", "PBM"

collection

Character: JASPAR matrix profile collection name. USE rba_jaspar_collections to get a list of collection names.

release

Numeric: (default = 2022) Which JASPAR database release to use? Available options are: 2014, 2016, 2018, 2020, and 2022.

only_last_version

Logical: (default = FALSE) If TRUE, only the latest version of a matrix profile will be returned.

order

Character: A character string or a vector of character strings of field names that will be used to order the results.
Providing multiple field names is supported. You can also use prefix "-" before a field name to indicate reverse ordering.

page_size

Numeric: (default = 1000) This resource returns paginated results. What is the maximum numbers of results that you want to retrieve per a page? Accepted values are between 1 and 1000.

page

Numeric: Which page of the results to retrieve? The accepted values depend on the page size and number of results.

...

rbioapi option(s). See rba_options's arguments manual for more information on available options.

Value

A list that contains a data frame of matrix profiles' information.

Details

Note that this is a search function. Thus, you are not required to fill every argument; You may use whatever combinations of arguments you see fit for your query.
The results are paginated. You can control the page's size number with the function's arguments. Also, you can use rba_pages to automatically iterate over multiple pages.

Corresponding API Resources

"GET "https://jaspar.genereg.net/api/v1/api/v1/matrix/"

References

Jaime A Castro-Mondragon, Rafael Riudavets-Puig, Ieva Rauluseviciute, Roza Berhanu Lemma, Laura Turchi, Romain Blanc-Mathieu, Jeremy Lucas, Paul Boddie, Aziz Khan, Nicolás Manosalva Pérez, Oriol Fornes, Tiffany Y Leung, Alejandro Aguirre, Fayrouz Hammal, Daniel Schmelter, Damir Baranasic, Benoit Ballester, Albin Sandelin, Boris Lenhard, Klaas Vandepoele, Wyeth W Wasserman, François Parcy, Anthony Mathelier, JASPAR 2022: the 9th release of the open-access database of transcription factor binding profiles, Nucleic Acids Research, Volume 50, Issue D1, 7 January 2022, Pages D165–D173, https://doi.org/10.1093/nar/gkab1113

JASPAR API Documentation

Citations note on JASPAR website

See also

Examples

# \donttest{
rba_jaspar_matrix_search(term = "FOX")
rba_jaspar_matrix_search(tf_name = "FOXP3")
rba_jaspar_matrix_search(tf_name = "FOXP3", only_last_version = TRUE)
rba_jaspar_matrix_search(tf_class = "Zipper-Type")
rba_jaspar_matrix_search(tax_group = "insects")
rba_jaspar_matrix_search(page_size = 100)
# }