Get information about JASPAR database releases

If a release number was supplied, this function will return the details of that release. Otherwise, if the function was called without "release" argument, a list of all JASPAR database releases will be returned.

rba_jaspar_releases(release_number = NULL, ...)

Arguments

release_number
Numeric: Which JASPAR database release number information's to retrieve? If left NULL (the default), a list of all JASPAR database releases will be returned. Available options are 1 to 8.

...
rbioapi option(s). See rba_options's arguments manual for more information on available options.

Value

A list that contains all JASPAR database releases' information or details of a particular release.

Corresponding API Resources

"GET "http://jaspar.genereg.net/api/v1/releases/" "GET "http://jaspar.genereg.net/api/v1/releases/release_number/"

References

Fornes O, Castro-Mondragon JA, Khan A, et al. JASPAR 2020: update of the open-access database of transcription factor binding profiles. Nucleic Acids Res. 2019; doi: 10.1093/nar/gkz1001

Khan, A. and Mathelier, A. JASPAR RESTful API: accessing JASPAR data from any programming language. Bioinformatics, 2017, doi: 10.1093/bioinformatics/btx804

Khan, A. et al. JASPAR 2018: update of the open-access database of transcription factor binding profiles and its web framework. Nucleic Acids Res. 2018; 46:D260–D266, doi: 10.1093/nar/gkx1126

JASPAR API Documentation

See also

Examples

# \donttest{
rba_jaspar_releases()
rba_jaspar_releases(7)
# }