Search TF flexible models (TFFMs) available in JASPAR

You can use this function to list the JASPAR TF flexible models (TFFMs) that match your search query, or run the function without any arguments to return a list of every matrix profile available in the latest release.

Usage

rba_jaspar_tffm_search(
  term = NULL,
  release = 2020,
  tax_group = NULL,
  search = NULL,
  order = NULL,
  page_size = 1000,
  page = 1,
  ...
)

Arguments

term

Character: A search term.

release

Numeric: (default = 2020) Which JASPAR database release to use? Available options are: 2014, 2016, 2018, and 2020.

tax_group

Character: Taxonomic group. Use rba_jaspar_taxons to get a list of supported Taxonomic groups.

search

Character: A search term.

order

Character: A character string or a vector of character strings of field names that will be used to order the results.
Providing multiple field names is supported. You can also use prefix "-" before a field name to indicate reverse ordering.

page_size

Numeric: (default = 1000) This resource returns paginated results. What is the maximum numbers of results that you want to retrieve per a page? Accepted values are between 1 and 1000.

page

Numeric: Which page of the results to retrieve? The accepted values depend on the page size and number of results.

...

rbioapi option(s). See rba_options's arguments manual for more information on available options.

Value

A list that contains a data frame with information of query hits' TFFMs.

Details

Note that this is a search function. Thus, you are not required to fill every argument; You may use whatever combinations of arguments you see fit for your query.
The results are paginated. You can control the page's size number with the function's arguments. Also, you can use rba_pages to automatically iterate over multiple pages.

Corresponding API Resources

"GET "http://jaspar.genereg.net/api/v1/api/v1/tffm/"

References

Fornes O, Castro-Mondragon JA, Khan A, et al. JASPAR 2020: update of the open-access database of transcription factor binding profiles. Nucleic Acids Res. 2019; doi: 10.1093/nar/gkz1001

Khan, A. and Mathelier, A. JASPAR RESTful API: accessing JASPAR data from any programming language. Bioinformatics, 2017, doi: 10.1093/bioinformatics/btx804

Khan, A. et al. JASPAR 2018: update of the open-access database of transcription factor binding profiles and its web framework. Nucleic Acids Res. 2018; 46:D260–D266, doi: 10.1093/nar/gkx1126

JASPAR API Documentation

See also

Examples

# \donttest{
rba_jaspar_tffm_search(term = "FOX")
rba_jaspar_tffm_search(tax_group = "insects")
rba_jaspar_tffm_search(page_size = 100)
# }