Get Supported Enrichment Categories for a Species and miRNA Type

For each Combination of species and miRNA type, Only a pre-defined categories groups are supported. Use this function to retrieve a list of supported categories for a given combination of Species and miRNA type.

Usage

rba_mieaa_cats(mirna_type, species, ...)

Arguments

mirna_type

Type of your miRNA accession. either "mature" or "precursor".

species

Fully or partially matching Scientific name, abbreviation or NCBI taxon ID of one of the following species:

"Homo sapiens", "hsa" or 9606

"Mus musculus", "mmu" or 10090

"Rattus norvegicus", "rno" or 10116

"Arabidopsis thaliana", "ath" or 3702

"Bos taurus", "bta" or 9913

"Caenorhabditis elegans", "cel" or 6239

"Drosophila melanogaster", "dme" or 7227

"Danio rerio", "dre" or 7955

"Gallus gallus", "gga" or 9031

"Sus scrofa", "ssc" or 9823

...

rbioapi option(s). See rba_options's arguments manual for more information on available options.

Value

a named character vector with the supported categories for your supplied input combination.

Corresponding API Resources

"GET https://ccb-compute2.cs.uni-saarland.de/mieaa2/api/v1/enrichment_categories/{species}/{mirna_type}/"

References

Fabian Kern, Tobias Fehlmann, Jeffrey Solomon, Louisa Schwed, Nadja Grammes, Christina Backes, Kendall Van Keuren-Jensen, David Wesley Craig,Eckart Meese, Andreas Keller, miEAA 2.0: integrating multi-species microRNA enrichment analysis and workflow management systems, Nucleic Acids Research, Volume 48, Issue W1, 02 July 2020, Pages W521–W528, https://doi.org/10.1093/nar/gkaa309

miEAA browsable API tutorial

Citations note on miEAA website

See also

Examples

# \donttest{
rba_mieaa_cats("mature", "Homo sapiens")
# }