Retrieve Results of a finished Enrichment Analysis from miEAA

After your submitted enrichment analysis request has finished (check using rba_mieaa_enrich_status), you can retrieve the results using this function.

Usage

rba_mieaa_enrich_results(job_id, sort_by = "p_adjusted", sort_asc = TRUE, ...)

Arguments

job_id

The job-id (a character string) of a submitted enrichment analysis.

sort_by

A column name to the result's table based on that. one of: "category", "subcategory", "enrichment", "p_value", "p_adjusted" (default), "q_value" or "observed" .

sort_asc

(logical) If TRUE, the results will be sorted in ascending order. If FALSE, the results will be sorted in descending order.

...

rbioapi option(s). See rba_options's arguments manual for more information on available options.

Value

A data frame with your enrichment analysis results.

Details

Note that using rba_mieaa_enrich is a more convenient way to automatically perform this and other required function calls to enrich your input miRNA-set using miEAA.

Corresponding API Resources

"GET https://ccb-compute2.cs.uni-saarland.de/mieaa2/api/v1/results/job_id"

References

Fabian Kern, Tobias Fehlmann, Jeffrey Solomon, Louisa Schwed, Nadja Grammes, Christina Backes, Kendall Van Keuren-Jensen, David Wesley Craig,Eckart Meese, Andreas Keller, miEAA 2.0: integrating multi-species microRNA enrichment analysis and workflow management systems, Nucleic Acids Research, Volume 48, Issue W1, 02 July 2020, Pages W521–W528, https://doi.org/10.1093/nar/gkaa309

miEAA browsable API tutorial

See also

Examples

if (FALSE) {
rba_mieaa_enrich_results("f52d1aef-6d3d-4d51-9020-82e68fe99012")
}