Map A Gene-set to PANTHER Database

Using this function, you can search your genes in PANTHER database and retrieve attributes and annotations associated to your genes.

rba_panther_mapping(genes, organism, ...)

Arguments

genes
Character vector of genes identifiers with maximum length of 1000. Can be any of: Ensemble gene ID, Ensemble protein ID, Ensemble transcript ID, Entrez gene ID, gene symbol, NCBI GI, HGNC ID, International protein index ID, NCBI UniGene ID, UniProt accession and/or UniProt ID.

organism
(numeric) NCBI taxon ID. run rba_panther_info with argument 'what = "organisms"' to get a list of PANTHER's supported organisms.

...
rbioapi option(s). Refer to rba_options's arguments documentation for more information on available options.

Value

A list containing your unmapped inputs and mapped genes with pertinent information.

Corresponding API Resources

"GET http://www.pantherdb.org/services/oai/pantherdb/geneinfo"

References

Mi, H., Muruganujan, A., Ebert, D., Huang, X., & Thomas, P. D. (2019). PANTHER version 14: more genomes, a new PANTHER GO-slim and improvements in enrichment analysis tools. Nucleic acids research, 47(D1), D419-D426.

Mi, H., Muruganujan, A., Huang, X., Ebert, D., Mills, C., Guo, X., & Thomas, P. D. (2019). Protocol Update for large-scale genome and gene function analysis with the PANTHER classification system (v. 14.0). Nature protocols, 14(3), 703-721.

PANTHER Services Details

See also

Examples

# \donttest{
rba_panther_mapping(genes = c("Cd40", 7124, "ENSG00000203747", "P33681"),
    organism = 9606)
# }