Search PANTHER for Orthologs of Gene(s)

Using this function you can search and retrieve orthologs of given gene(s), and optionally return the corresponding position in the target organisms' protein sequences.

rba_panther_ortholog(
  genes,
  organism,
  type = "all",
  target_organisms = NA,
  seq_pos = NA,
  include_msa = NA,
  ...
)

Arguments

genes
Character vector of genes identifiers with maximum length of 10 or only one if seq_pos is provided. Can be any of: Ensemble gene ID, Ensemble protein ID, Ensemble transcript ID, Entrez gene ID, gene symbol, NCBI GI, HGNC ID, International protein index ID, NCBI UniGene ID, UniProt accession and/or UniProt ID.

organism
(numeric) NCBI taxon ID of the organism of your provided genes. run rba_panther_info with argument 'what = "organisms"' to get a list of PANTHER's supported organisms.

type
Ortholog types to return. either "all" (default) or "LDO" to only return least diverged orthologs.

target_organisms
(numeric) NCBI taxon ID(s) to filter the results. run rba_panther_info with argument 'what = "organisms"' to get a list of PANTHER's supported organisms.

seq_pos
(Numeric) A position in the protein's sequence of the provided gene. should be in the range of the protein's length.

include_msa
(Logical) Only if a sequence position is provided, should MSA (Multiple Sequence Alignment) information be included in the results?

...
rbioapi option(s). Refer to rba_options's arguments documentation for more information on available options.

Value

A data frame with Orthologs information.

Corresponding API Resources

"POST http://www.pantherdb.org/services/oai/pantherdb/ortholog/matchortho"
"POST http://www.pantherdb.org/services/oai/pantherdb/ortholog/homologpos"

References

Mi, H., Muruganujan, A., Ebert, D., Huang, X., & Thomas, P. D. (2019). PANTHER version 14: more genomes, a new PANTHER GO-slim and improvements in enrichment analysis tools. Nucleic acids research, 47(D1), D419-D426.

Mi, H., Muruganujan, A., Huang, X., Ebert, D., Mills, C., Guo, X., & Thomas, P. D. (2019). Protocol Update for large-scale genome and gene function analysis with the PANTHER classification system (v. 14.0). Nature protocols, 14(3), 703-721.

PANTHER Services Details

See also

Examples

# \donttest{
rba_panther_ortholog("CD40", organism = 9606, type = "LDO")
# }