Maps Molecule Identifiers

Use this function to map molecule identifiers of different species to Reactome Identifiers.

rba_reactome_analysis_mapping(
  input,
  input_format = NA,
  projection = FALSE,
  interactors = FALSE,
  ...
)

Arguments

input
A vector, local file path or URL that points to your identifiers list.

input_format
(Optional) This function will automatically identify your provided input's format. But in case of unexpected issues or if you want to be explicit, set this argument to one of:

"vector": If you provided a simple vector (numeric or character) as input.

"file": If you provided a local file path pointing to a correctly-formatted text file.

"url": If you provided a URL pointing to a correctly-formatted text file.

projection
Logical (default = FALSE) Should non-human identifiers be projected to their human equivalents? (using Reactome orthology data)

interactors
Logical (default = FALSE) Should IntAct interaction data be included?

...
rbioapi option(s). Refer to rba_options's arguments documentation for more information on available options.

Value

List containing your identifiers and the IDS and resources they are mapped to.

Corresponding API Resources

"GET https://reactome.org/AnalysisService/mapping"
"GET https://reactome.org/AnalysisService/mapping/form"
"GET https://reactome.org/AnalysisService/mapping/form/projection"
"GET https://reactome.org/AnalysisService/mapping"
"GET https://reactome.org/AnalysisService/mapping/url"
"GET https://reactome.org/AnalysisService/mapping/url/projection"

References

Fabregat A, Sidiropoulos K, Viteri G, Forner O, Marin-Garcia P, Arnau V, D'Eustachio P, Stein L, Hermjakob H. Reactome pathway analysis: a high-performance in-memory approach. BMC bioinformatics. 2017 Mar;18(1) 142. doi: 10.1186/s12859-017-1559-2. PubMed PMID: 28249561. PubMed Central PMCID: PMC5333408.

Reactome Analysis Services API Documentation

See also

Examples

# \donttest{
rba_reactome_analysis_mapping(c("Q8SQ34", "cd40"))
# }