Generate PDF file with Reactome Analysis Results

Use this function to save a detailed report of your previous analysis (That you have done with rba_reactome_analysis). You need to provide a 'token' associated to your previous analysis.

rba_reactome_analysis_pdf(
  token,
  species,
  save_to = NA,
  number = 25,
  resource = "TOTAL",
  diagram_profile = "Modern",
  analysis_profile = "Standard",
  fireworks_profile = "Barium Lithium",
  ...
)

Arguments

token
A token associated to your previous Reactome analysis.

species
Numeric or Character: NCBI Taxonomy identifier (Human Taxonomy ID is 9606.) or species name (e.g. "Homo sapiens"). Refer to rba_reactome_species or Reactome Data Schema: Entries: Species.

save_to
NA or Character:

NA: Save the file to an automatically-generated path.

Character string: A valid file path to save the file to.

number
Numeric: Maximum number of the reported pathways. Cannot not be greater than 50.

resource
Filter results based on the resource. Default is "TOTAL", available choices are: "TOTAL", "UNIPROT", "ENSEMBL", "CHEBI", "IUPHAR", "MIRBASE", "NCBI_PROTEIN", "EMBL", "COMPOUND", "ENTITIES_FDR" or "PUBCHEM_COMPOUND".

diagram_profile
Color profile of diagrams, should be either "Modern" (default) or "Standard".

analysis_profile
Color profile of analysis, should be one of: "Standard" (default), "Strosobar" or "Copper Plus".

fireworks_profile
Color profile of overview diagram, should be one of: "Copper", "Copper Plus", "Barium Lithium" or "calcium salts".

...
rbioapi option(s). Refer to rba_options's arguments documentation for more information on available options.

Value

NULL, a PDF file will be saved to disk.

Details

Token is associated to each Reactome analysis results and kept by Reactome for at least 7 days. You can locate it in rba_reactome_analysis's output, under a sub-list named "summary" (i.e. results$summary$token).
Note that Reactome will store your token for only 7 days. You can download your full results with rba_reactome_analysis_download, and re-import it anytime to reactome (using rba_reactome_analysis_import) to generate a new token. Use rba_reactome_analysis_download to save your results in other formats.

Corresponding API Resources

"GET https://reactome.org/AnalysisService/report/token/species/ filename.pdf"

References

Fabregat A, Sidiropoulos K, Viteri G, Forner O, Marin-Garcia P, Arnau V, D'Eustachio P, Stein L, Hermjakob H. Reactome pathway analysis: a high-performance in-memory approach. BMC bioinformatics. 2017 Mar;18(1) 142. doi: 10.1186/s12859-017-1559-2. PubMed PMID: 28249561. PubMed Central PMCID: PMC5333408.

Reactome Analysis Services API Documentation

See also

Examples

if (FALSE) {
rba_reactome_analysis_pdf(token = "MjAyMDEwMTYwMTI3MTNfMjY1MjM%3D",
    species = 9606, save_to = "my_analysis.pdf")
}