Get Complexes That Include a Molecule

This function will retrieve a list of complexes that include your supplied molecule as a component.

Usage

rba_reactome_complex_list(id, resource, ...)

Arguments

id

Molecule's external Identifier

resource

What is the resource of your supplied ID? see: Reactome External Identifiers

...

rbioapi option(s). See rba_options's arguments manual for more information on available options.

Value

Data frame where each row is a complex containing your supplied molecule and columns are pertinent information.

Corresponding API Resources

"GET https://reactome.org/ContentService/data/complexes/{resource}/ {identifier}"

References

Marc Gillespie, Bijay Jassal, Ralf Stephan, Marija Milacic, Karen Rothfels, Andrea Senff-Ribeiro, Johannes Griss, Cristoffer Sevilla, Lisa Matthews, Chuqiao Gong, Chuan Deng, Thawfeek Varusai, Eliot Ragueneau, Yusra Haider, Bruce May, Veronica Shamovsky, Joel Weiser, Timothy Brunson, Nasim Sanati, Liam Beckman, Xiang Shao, Antonio Fabregat, Konstantinos Sidiropoulos, Julieth Murillo, Guilherme Viteri, Justin Cook, Solomon Shorser, Gary Bader, Emek Demir, Chris Sander, Robin Haw, Guanming Wu, Lincoln Stein, Henning Hermjakob, Peter D’Eustachio, The reactome pathway knowledgebase 2022, Nucleic Acids Research, 2021;, kab1028, https://doi.org/10.1093/nar/gkab1028

Griss J, Viteri G, Sidiropoulos K, Nguyen V, Fabregat A, Hermjakob H. ReactomeGSA - Efficient Multi-Omics Comparative Pathway Analysis. Mol Cell Proteomics. 2020 Sep 9. doi: 10.1074/mcp. PubMed PMID: 32907876.

Reactome Content Services API Documentation

Citations note on Reactome website

See also

Examples

# \donttest{
rba_reactome_complex_list(id = "3845", resource = "NCBI Gene")
# }
# \donttest{
rba_reactome_complex_list(id = "P00533", resource = "UniProt")
# }