Exports A Reactome Event to SBGN or SBML

This function will export a supplied Reactome Event (Pathway or Reaction) to a SBGN (Systems Biology Graphical Notation) or SBML (Systems Biology Markup Language)

rba_reactome_exporter_event(event_id, output_format, save_to = NULL, ...)

Arguments

event_id
Reactome event's database IDs (DbId) or Stable IDs (StId).

output_format
Either "sbgn" or "sbml".

save_to
NULL or Character:

NULL: Save the file to an automatically-generated path.

Character string: A valid file path to save the file to.

...
rbioapi option(s). See rba_options's arguments manual for more information on available options.

Value

NULL, According to the inputs, a SBGN or SBML file will be saved to disk.

Corresponding API Resources

"GET https://reactome.org/ContentService//exporter/event/ identifier.sbgn" "GET https://reactome.org/ContentService//exporter/event/ identifier.sbml"

References

Jassal B, Matthews L, Viteri G, Gong C, Lorente P, Fabregat A, Sidiropoulos K, Cook J, Gillespie M, Haw R, Loney F, May B, Milacic M, Rothfels K, Sevilla C, Shamovsky V, Shorser S, Varusai T, Weiser J, Wu G, Stein L, Hermjakob H, D'Eustachio P. The reactome pathway knowledgebase. Nucleic Acids Res. 2020 Jan 8;48(D1):D498-D503. doi: 10.1093/nar/gkz1031. PubMed PMID: 31691815.

Reactome Content Services API Documentation

See also

Examples

if (FALSE) {
rba_reactome_exporter_event(event_id = "R-HSA-177929",
    output_format = "sbgn",
    save_to = "R-HSA-177929.sbgn")
}
if (FALSE) {
rba_reactome_exporter_event(event_id = "R-HSA-177929",
    output_format = "sbgn")
}