Get Orthologous (Computationally Inferred) Events

Reactome incorporate manually curated human reactions and PANTHER's protein homology data to Computationally infer events in other eukaryotic species.

rba_reactome_orthology(event_ids, species_dbid, ...)

Arguments

event_ids
Human Reactome event ID(s) to retrieve their orthologous events.

species_dbid
Reactome database ID (DbId) of the target species. (e.g Mus musculus is 48892). See rba_reactome_species or Reactome Data Schema: Entries: Species.

...
rbioapi option(s). See rba_options's arguments manual for more information on available options.

Value

List containing found Orthologous event(s) in your supplied species and their pertinent information.

Details

In version 73 (11 June 2020), using an orthology-based approach, Homo sapiens events was projected to 18,654 orthologous pathways (with 81,835 orthologous proteins) in 15 non-human species.
See Reactome Computationally Inferred Events for more information.

Corresponding API Resources

"POST https://reactome.org/ContentService/data/orthologies/ids/ species/speciesId"

References

Jassal B, Matthews L, Viteri G, Gong C, Lorente P, Fabregat A, Sidiropoulos K, Cook J, Gillespie M, Haw R, Loney F, May B, Milacic M, Rothfels K, Sevilla C, Shamovsky V, Shorser S, Varusai T, Weiser J, Wu G, Stein L, Hermjakob H, D'Eustachio P. The reactome pathway knowledgebase. Nucleic Acids Res. 2020 Jan 8;48(D1):D498-D503. doi: 10.1093/nar/gkz1031. PubMed PMID: 31691815.

Reactome Content Services API Documentation

See also

Examples

# \donttest{
rba_reactome_orthology(event_ids = c("R-HSA-6799198", " R-HSA-72764"),
    species_dbid = 49633)
# }