Get lower level pathways Containing a 'Physical Entity' or Event

Use this function to search the event hierarchy and retrieve a list of all lower level pathways (non TopLevelPathway class) that contain a given 'Physical Entity' or Event. See "Arguments section" on how to modify your search.

rba_reactome_pathways_low(
  entity_id,
  with_diagram = FALSE,
  all_forms = FALSE,
  species = NULL,
  ...
)

Arguments

entity_id
The entity that should exist in the pathways.

with_diagram
Logical: only include pathways with diagram?

all_forms
Logical: should other variants of your supplied entity_id be considered? (e.g. same molecule but in different compartment, secretory form etc.) see rba_reactome_participants's "Details section" to learn more about how Reactome classifies molecules.

species
(optional) Numeric or Character: confine your search to a specific species by providing it's NCBI Taxonomy identifier (Human Taxonomy ID is 9606) or species name (e.g. "Homo sapiens"). See rba_reactome_species or Reactome Data Schema: Entries: Species.

...
rbioapi option(s). See rba_options's arguments manual for more information on available options.

Value

Data frame where each row is a pathway that contains your supplied entity and columns are pertinent information.

Corresponding API Resources

"GET https://reactome.org/ContentService/data/pathways/low/entity/id"
"GET https://reactome.org/ContentService/data/pathways/low/diagram/ entity/id"
"GET https://reactome.org/ContentService/data/pathways/low/diagram/ entity/id/allForms"

References

Jassal B, Matthews L, Viteri G, Gong C, Lorente P, Fabregat A, Sidiropoulos K, Cook J, Gillespie M, Haw R, Loney F, May B, Milacic M, Rothfels K, Sevilla C, Shamovsky V, Shorser S, Varusai T, Weiser J, Wu G, Stein L, Hermjakob H, D'Eustachio P. The reactome pathway knowledgebase. Nucleic Acids Res. 2020 Jan 8;48(D1):D498-D503. doi: 10.1093/nar/gkz1031. PubMed PMID: 31691815.

Reactome Content Services API Documentation

See also

Examples

# \donttest{
rba_reactome_pathways_low(entity_id = "R-HSA-199420")
# }
# \donttest{
rba_reactome_pathways_low(entity_id = "R-HSA-199420", with_diagram = TRUE)
# }
# \donttest{
rba_reactome_pathways_low(entity_id = "R-HSA-199420", with_diagram = TRUE,
    all_forms = TRUE)
# }