Get Protein-Protein Interaction Enrichment

Even when there is no annotation for your input proteins, STRING can Compare your Given proteins interactions pattern with the background proteome-wide interaction distribution to determine if your given set of proteins are functionally related.

rba_string_enrichment_ppi(ids, species = NA, required_score = NA, ...)

Arguments

ids
Your protein ID(s). It is strongly recommended to provide STRING IDs. See rba_string_map_ids for more information.

species
Numeric: NCBI Taxonomy identifier; Human Taxonomy ID is 9606. (Recommended, but optional if your input is less than 100 IDs.)

required_score
Numeric: A minimum of interaction score for an interaction to be included in the image. if not provided, the threshold will be applied by STRING Based in the network. (low Confidence = 150, Medium Confidence = 400, High Confidence = 700, Highest confidence = 900)

...
rbioapi option(s). Refer to rba_options's arguments documentation for more information on available options.

Value

A list with protein-protein interaction enrichment results.

Corresponding API Resources

"POST https://string-db.org/api/[output_format]/ppi_enrichment?identifiers= [your_identifiers]&[optional_parameters]"

References

Szklarczyk D, Gable AL, Lyon D, Junge A, Wyder S, Huerta-Cepas J, Simonovic M, Doncheva NT, Morris JH, Bork P, Jensen LJ, Mering CV. STRING v11: protein-protein association networks with increased coverage, supporting functional discovery in genome-wide experimental datasets. Nucleic Acids Res. 2019 Jan 8;47(D1):D607-D613. doi: 10.1093/nar/gky1131. PMID: 30476243; PMCID: PMC6323986.

STRING API Documentation

See also

Examples

# \donttest{
rba_string_enrichment_ppi(ids = c("p53", "BRCA1", "cdk2", "Q99835",
       "CDC42", "CDK1", "KIF23", "PLK1", "RAC2", "RACGAP1"),
    species = 9606)
# }