UniProt maintains sequence annotations (features) that describe regions in the protein sequence. Using this function, you can search and retrieve UniProt proteins' sequence annotations (features). you may also refine your search query with variety of modifiers.

Note that this is a search function. Thus, you are not required to fill every argument; You may use whatever combinations of arguments you see fit for your query.
UniProt Entries are grouped in two sections:

Reviewed(Swiss-Prot): Manually annotated records with information extracted from literature and curator-evaluated computational analysis.

Unreviewed (TrEMBL): Computationally analyzed records that await full manual annotation.

rba_uniprot_features_search(
  accession = NULL,
  gene = NULL,
  exact_gene = NULL,
  protein = NULL,
  reviewed = NULL,
  organism = NULL,
  taxid = NULL,
  categories = NULL,
  types = NULL,
  ...
)

Arguments

accession
UniProtKB primary or secondary accession(s). You can provide up to 100 accession numbers.

gene
UniProt gene name(s). You can provide up to 20 gene names. e.g. if you provide "CD40", "CD40 ligand" will also be included.

exact_gene
UniProt exact gene name(s). You can provide up to 20 exact gene names. e.g. if you provide "CD40", "CD40 ligand" will not be included in the results.

protein
UniProt protein name

reviewed
Logical: If TRUE, only return "UniProtKB/Swiss-Prot" (reviewed) entries; If FALSE, only return TrEMBL (un-reviewed) entries.

organism
Organism name.

taxid
NIH-NCBI Taxon ID. You can provide up to 20 taxon IDs.

categories
Sequence annotation (Features) categories (subsection). accepted values are: "MOLECULE_PROCESSING", "TOPOLOGY", "SEQUENCE_INFORMATION", "STRUCTURAL", "DOMAINS_AND_SITES", "PTM", "VARIANTS" and/or "MUTAGENESIS". You can provide up to 8 categories.

types
Sequence annotation (Features) types. accepted values are: "INIT_MET", "SIGNAL", "PROPEP", "TRANSIT", "CHAIN", "PEPTIDE", "TOPO_DOM", "TRANSMEM", "DOMAIN", "REPEAT", "CA_BIND", "ZN_FING", "DNA_BIND", "NP_BIND", "REGION", "COILED", "MOTIF", "COMPBIAS", "ACT_SITE", "METAL", "BINDING", "SITE", "NON_STD", "MOD_RES", "LIPID", "CARBOHYD", "DISULFID", "CROSSLNK", "VAR_SEQ", "VARIANT", "MUTAGEN", "UNSURE", "CONFLICT", "NON_CONS", "NON_TER", "HELIX", "TURN", "STRAND" and/or "INTRAMEM". You can provide up to 20 types.

...
rbioapi option(s). Refer to rba_options's arguments documentation for more information on available options.

Value

List where each element corresponds to one UniProt entity returned by your search query. The element itself is a sub-list containing all information that UniProt has about that entity.

Corresponding API Resources

"GET https://www.ebi.ac.uk/proteins/api/features"

References

Andrew Nightingale, Ricardo Antunes, Emanuele Alpi, Borisas Bursteinas, Leonardo Gonzales, Wudong Liu, Jie Luo, Guoying Qi, Edd Turner, Maria Martin, The Proteins API: accessing key integrated protein and genome information, Nucleic Acids Research, Volume 45, Issue W1, 3 July 2017, Pages W539–W544, https://doi.org/10.1093/nar/gkx237

Proteins API Documentation

See also

Examples

# \donttest{
rba_uniprot_features_search(accession = "Q99616")
# }
# \donttest{
rba_uniprot_features_search(gene = "cd40")
# }
# \donttest{
rba_uniprot_features_search(gene = "cd40 ligand")
# }
# \donttest{
rba_uniprot_features_search(gene = "cd40",  reviewed = TRUE)
# }
# \donttest{
rba_uniprot_features_search(accession = "Q99616",
    categories = c("MOLECULE_PROCESSING", "TOPOLOGY"))
# }
# \donttest{
rba_uniprot_features_search(accession = "Q99616", types = "DISULFID")
# }