Search Gene-Centric Proteins

Using this function you can search UniProt for available gene-centrics from proteomes. For more information, see What are proteomes? and Automatic gene-centric isoform mapping for eukaryotic reference proteome entries. You may also refine your search with modifiers upid, accession and gene. See "Arguments section" for more information.

Usage

rba_uniprot_genecentric_search(upid = NULL, accession = NULL, gene = NULL, ...)

Arguments

upid

UniProt Proteome identifier (UPID). You can supply up to 100 UPIDs.

accession

UniProtKB primary or secondary accession(s). You can supply up to 100 accession numbers.

gene

unique gene identifier(s) found in MOD, Ensembl, Ensembl Genomes, OLN, ORF or UniProt Gene Name.

...

rbioapi option(s). See rba_options's arguments manual for more information on available options.

Value

a list containing gene-centric proteins search hits.

Details

Note that this is a search function. Thus, you are not required to fill every argument; You may use whatever combinations of arguments you see fit for your query.

Corresponding API Resources

"GET https://ebi.ac.uk/proteins/api/genecentric"

References

The UniProt Consortium, UniProt: the universal protein knowledgebase in 2021, Nucleic Acids Research, Volume 49, Issue D1, 8 January 2021, Pages D480–D489, https://doi.org/10.1093/nar/gkaa1100

Andrew Nightingale, Ricardo Antunes, Emanuele Alpi, Borisas Bursteinas, Leonardo Gonzales, Wudong Liu, Jie Luo, Guoying Qi, Edd Turner, Maria Martin, The Proteins API: accessing key integrated protein and genome information, Nucleic Acids Research, Volume 45, Issue W1, 3 July 2017, Pages W539–W544, https://doi.org/10.1093/nar/gkx237

Proteins API Documentation

Citations note on UniProt website

See also

Examples

# \donttest{
rba_uniprot_genecentric_search(accession = "P59594")
# }
# \donttest{
rba_uniprot_genecentric_search(gene = "Spike")
# }
# \donttest{
rba_uniprot_genecentric_search(upid = "UP000000354")
# }