Get proteome by proteome/proteins UPID

UniProt collects and annotates proteomes(Protein sets expressed in an organism). Using this function you can search UniProt for available proteomes. see What are proteomes? for more information.

rba_uniprot_proteomes(upid, get_proteins = FALSE, reviewed = NA, ...)

Arguments

upid
UniProt Proteome identifier (UPID). You can provide up to 100 UPIDs.

get_proteins
logical: set FALSE (default) to only return information of the proteome with provided UPID, set TRUE to also return the proteins of the provided proteome UPID.

reviewed
Logical: Only considered when get_proteins is TRUE. If TRUE, only return "UniProtKB/Swiss-Prot" (reviewed) proteins; If FALSE, only return TrEMBL (un-reviewed) entries. leave it as NA if you do not want to filter proteins based on their review status.

...
rbioapi option(s). Refer to rba_options's arguments documentation for more information on available options.

Value

a list containing information of the proteome with your provided UPID that can contain the proteomes protein entries based on the value of get_proteins argument.

Corresponding API Resources

"GET https://ebi.ac.uk/proteins/api/proteomes/proteins/upid"
"GET https://ebi.ac.uk/proteins/api/proteomes/upid"

References

Andrew Nightingale, Ricardo Antunes, Emanuele Alpi, Borisas Bursteinas, Leonardo Gonzales, Wudong Liu, Jie Luo, Guoying Qi, Edd Turner, Maria Martin, The Proteins API: accessing key integrated protein and genome information, Nucleic Acids Research, Volume 45, Issue W1, 3 July 2017, Pages W539–W544, https://doi.org/10.1093/nar/gkx237

Proteins API Documentation

See also

Examples

# \donttest{
rba_uniprot_proteomes(upid = "UP000000354")
# }
# \donttest{
rba_uniprot_proteomes(upid = "UP000000354", get_proteins = TRUE)
# }