Search Proteomes in UniProt

UniProt collects and annotates proteomes (Protein sets expressed in an organism). Using this function you can search UniProt for available proteomes. see What are proteomes? for more information. You may also refine your search with modifiers such as keyword, taxon id etc. refer to "Arguments section" for more information.

rba_uniprot_proteomes_search(
  name = NA,
  upid = NA,
  taxid = NA,
  keyword = NA,
  xref = NA,
  genome_acc = NA,
  is_ref_proteome = NA,
  is_redundant = NA,
  ...
)

Arguments

name
a keyword in proteome's name

upid
UniProt Proteome identifier (UPID). You can provide up to 100 UPIDs.

taxid
NIH-NCBI Taxon ID. You can provide up to 20 taxon IDs.

keyword
Limit the search to entries that contain your provided keyword. see: UniProt Keywords

xref
Proteome cross-references such as Genome assembly ID or Biosample ID. You can provide up to 20 cross-reference IDs.

genome_acc
Genome accession associated with the proteome's components.

is_ref_proteome
(logical) If TRUE, only return reference proteomes; If FALSE, only returns non-reference proteomes; If NA (default), the results will not be filtered by this criteria see 'What are reference proteomes?' for more information.

is_redundant
(logical) If TRUE, only return redundant proteomes; If FALSE, only returns non-redundant proteomes; If NA (default), the results will not be filtered by redundancy. see 'Reducing proteome redundancy' for more information.

...
rbioapi option(s). Refer to rba_options's arguments documentation for more information on available options.

Value

A list where each element is a list that corresponds to a single proteome (search hit) and contains informations pertinent to that proteome.

Details

Note that this is a search function. Thus, you are not required to fill every argument; You may use whatever combinations of arguments you see fit for your query.

Corresponding API Resources

"GET https://ebi.ac.uk/proteins/api/proteomes"

References

Andrew Nightingale, Ricardo Antunes, Emanuele Alpi, Borisas Bursteinas, Leonardo Gonzales, Wudong Liu, Jie Luo, Guoying Qi, Edd Turner, Maria Martin, The Proteins API: accessing key integrated protein and genome information, Nucleic Acids Research, Volume 45, Issue W1, 3 July 2017, Pages W539–W544, https://doi.org/10.1093/nar/gkx237

Proteins API Documentation

See also

Examples

# \donttest{
rba_uniprot_proteomes_search(name = "SARS-CoV")
# }
# \donttest{
rba_uniprot_proteomes_search(name = "SARS-CoV", is_ref_proteome = TRUE)
# }
# \donttest{
rba_uniprot_proteomes_search(name = "SARS-CoV", is_ref_proteome = TRUE)
# }
# \donttest{
rba_uniprot_proteomes_search(genome_acc = "AY274119")
# }