Traverse UniProt Taxonomic Tree Path

Using this function you can retrieve nodes that are located in the top or the bottom of your supplied node in UniProt Taxonomy database tree

Usage

rba_uniprot_taxonomy_path(id, direction, depth = 5, ...)

Arguments

id

(numeric) a NCBI taxonomic identifier

direction

direction of the taxonomic path, either "TOP" or "BOTTOM".

depth

(numeric) How many levels should be traversed on the taxonomic tree? (from 1 to 5, default = 5)

...

rbioapi option(s). See rba_options's arguments manual for more information on available options.

Value

a nested list containing the node which are in the path specified by your supplied argument in the UniProt taxonomic tree.

Corresponding API Resources

"GET https://ebi.ac.uk/proteins/api/path"

References

The UniProt Consortium, UniProt: the universal protein knowledgebase in 2021, Nucleic Acids Research, Volume 49, Issue D1, 8 January 2021, Pages D480–D489, https://doi.org/10.1093/nar/gkaa1100

Andrew Nightingale, Ricardo Antunes, Emanuele Alpi, Borisas Bursteinas, Leonardo Gonzales, Wudong Liu, Jie Luo, Guoying Qi, Edd Turner, Maria Martin, The Proteins API: accessing key integrated protein and genome information, Nucleic Acids Research, Volume 45, Issue W1, 3 July 2017, Pages W539–W544, https://doi.org/10.1093/nar/gkx237

Proteins API Documentation

Citations note on UniProt website

See also

Examples

# \donttest{
rba_uniprot_taxonomy_path(id = 9606, direction = "TOP", depth = 3)
# }
# \donttest{
rba_uniprot_taxonomy_path(id = 207598, direction = "BOTTOM", depth = 3)
# }