rbioapi: User-Friendly R Interface to Biologic Web Services' API

Currently fully supports Enrichr, JASPAR, miEAA, PANTHER, Reactome, STRING, and UniProt! The goal of rbioapi is to provide a user-friendly and consistent interface to biological databases and services. In a way that insulates the user from the technicalities of using web services API and creates a unified and easy-to-use interface to biological and medical web services. This is an ongoing project; New databases and services will be added periodically. Feel free to suggest any databases or services you often use.

Supported Database/Services

Enrichr

JASPAR

MiEAA

PANTHER

Reactome

STRING

UniProt

See also

Author

Maintainer: Moosa Rezwani moosa.rezwani@gmail.com (ORCID)