Retrieve a List of available libraries from Enrichr

This function will retrieve a list of available libraries in Enrichr with their statistics. And by default, will save those names as a global option ("rba_enrichr_libs") to be available for other Enrichr functions that internally require the names of Enrichr libraries.

Usage

rba_enrichr_libs(store_in_options = FALSE, organism = "human", ...)

Arguments

store_in_options

logical: (default = TRUE) Should a list of available Enrichr libraries be saved as a global option?

organism

(default = "human") Which model organism version of Enrichr to use? Available options are: "human", (H. sapiens & M. musculus), "fly" (D. melanogaster), "yeast" (S. cerevisiae), "worm" (C. elegans) and "fish" (D. rerio).

...

rbioapi option(s). See rba_options's arguments manual for more information on available options.

Value

A data frame with the names of available library in Enrichr and their statistics.

Details

You should call this function once per R session with the argument 'store_in_options = TRUE' before using rba_enrichr_enrich or rba_enrichr.
Nevertheless, rbioapi will do this for you in the background at the first time you call any function which requires this.
Note that using rba_enrichr is a more convenient way to automatically perform this and other required function calls to enrich your input gene-set.

Corresponding API Resources

"GET https://maayanlab.cloud/Enrichr/datasetStatistics"

References

Chen, E.Y., Tan, C.M., Kou, Y. et al. Enrichr: interactive and collaborative HTML5 gene list enrichment analysis tool. Bioinformatics 14, 128 (2013). https://doi.org/10.1186/1471-2105-14-128

Maxim V. Kuleshov, Matthew R. Jones, Andrew D. Rouillard, Nicolas F. Fernandez, Qiaonan Duan, Zichen Wang, Simon Koplev, Sherry L. Jenkins, Kathleen M. Jagodnik, Alexander Lachmann, Michael G. McDermott, Caroline D. Monteiro, Gregory W. Gundersen, Avi Ma’ayan, Enrichr: a comprehensive gene set enrichment analysis web server 2016 update, Nucleic Acids Research, Volume 44, Issue W1, 8 July 2016, Pages W90–W97, https://doi.org/10.1093/nar/gkw377

Xie, Z., Bailey, A., Kuleshov, M. V., Clarke, D. J. B., Evangelista, J. E., Jenkins, S. L., Lachmann, A., Wojciechowicz, M. L., Kropiwnicki, E., Jagodnik, K. M., Jeon, M., & Ma’ayan, A. (2021). Gene set knowledge discovery with Enrichr. Current Protocols, 1, e90. doi: 10.1002/cpz1.90

Enrichr API Documentation

Citations note on Enrichr website

See also

Examples

# \donttest{
rba_enrichr_libs()
# }