Convert Between Mature and precursor miRNA Accession

miRBase miRNA accession could refer to either mature or precursor miRNAs. (see: miRNA naming conventions). Use this function to mature miRNA accession to corresponding miRNA accessions or vice versa.

rba_mieaa_convert_type(
  mirna,
  input_type,
  only_unique = FALSE,
  simple_output = FALSE,
  ...
)

Arguments

mirna
A vector of miRNA accessions to be converted.

input_type
Type of your provided miRNA accession. either "mature" or "precursor".

only_unique
(logical) miRBase precursor and mature miRNA accessions are not uniquely mapped. (i.e. you may get more than one results for a given accession). set this to TRUE to only retrieve the unique mappings. (default = FALSE)

simple_output
(logical) If FALSE (default), the result will be a two-columned data frame with your input and output accessions. Otherwise, if TRUE, only the output miRNA accessions will be returned.

...
rbioapi option(s). Refer to rba_options's arguments documentation for more information on available options.

Value

Depending on the arguments, a data frame or a character vectors containing the miRNA accessions in your output version.

Corresponding API Resources

"POST https://ccb-compute2.cs.uni-saarland.de/mieaa2/api/v1/mirna_precursor_converter/"

References

Fabian Kern, Tobias Fehlmann, Jeffrey Solomon, Louisa Schwed, Nadja Grammes, Christina Backes, Kendall Van Keuren-Jensen, David Wesley Craig,Eckart Meese, Andreas Keller, miEAA 2.0: integrating multi-species microRNA enrichment analysis and workflow management systems, Nucleic Acids Research, Volume 48, Issue W1, 02 July 2020, Pages W521–W528, https://doi.org/10.1093/nar/gkaa309

miEAA browsable API tutorial

See also

Examples

# \donttest{
Sys.sleep(1) # to prevent 429 error during R CMD check
rba_mieaa_convert_type(mirna = c("hsa-miR-20b-5p", "hsa-miR-144-5p"),
    input_type = "mature")
# }