PANTHER Tree Grafter Use this function to retrieve a PANTHER family's tree topology information with a node corresponding to your sequence grafted in the best location in that tree.

For more information, see: TreeGrafter: phylogenetic tree-based annotation of proteins with Gene Ontology terms and other annotations

rba_panther_tree_grafter(protein_seq, target_organisms = NULL, ...)

Arguments

protein_seq
A character string with the protein's sequence. Maximum allowed sequence length is 50kb.

target_organisms
(numeric) NCBI taxon ID(s) to filter the results. run rba_panther_info with argument 'what = "organisms"' to get a list of PANTHER's supported organisms.

...
rbioapi option(s). See rba_options's arguments manual for more information on available options.

Value

A list containing PANTHER tree topology information.

Corresponding API Resources

"GET http://www.pantherdb.org/services/oai/pantherdb/graftsequence"

References

Mi, H., Muruganujan, A., Ebert, D., Huang, X., & Thomas, P. D. (2019). PANTHER version 14: more genomes, a new PANTHER GO-slim and improvements in enrichment analysis tools. Nucleic acids research, 47(D1), D419-D426.

Mi, H., Muruganujan, A., Huang, X., Ebert, D., Mills, C., Guo, X., & Thomas, P. D. (2019). Protocol Update for large-scale genome and gene function analysis with the PANTHER classification system (v. 14.0). Nature protocols, 14(3), 703-721.

PANTHER Services Details

See also

Examples

# \donttest{
rba_panther_tree_grafter("MKVLWAALLVTFLAGCQAKVEQAVETE")
# }