Query and Retrieve any Reactome knowledge-base Object

Using this Comprehensive function, You can Retrieve any object from Reactome knowledge-base

rba_reactome_query(
  ids,
  enhanced = FALSE,
  map = FALSE,
  attribute_name = NULL,
  ...
)

Arguments

ids
A single or Multiple database IDs (DbId), Stable IDs (StId) or a mixture of both.

enhanced
Logical: (Default = FALSE) If 'TRUE' more information on the supplied entry will be returned. (You can set this argument to 'TRUE' Only when you supply a single ID).

map
(Default = FALSE) Should the supplied IDs be mapped? This argument will only be considered when you supply multiple IDs. (e.g. when you supply previous version of stable identifiers.)

attribute_name
(Optional) Only Return an Attribute of the supplied Database Object. (You can use this argument Only when you supply a single ID)

...
rbioapi option(s). See rba_options's arguments manual for more information on available options.

Value

List containing your query outputs.

Corresponding API Resources

"POST https://reactome.org/ContentService/data/query/ids"
"POST https://reactome.org/ContentService/data/query/ids/map"
"GET https://reactome.org/ContentService/data/query/id"
s"GET https://reactome.org/ContentService//data/query/enhanced/id"
"GET https://reactome.org/ContentService/data/query/id/attributeName"

References

Jassal B, Matthews L, Viteri G, Gong C, Lorente P, Fabregat A, Sidiropoulos K, Cook J, Gillespie M, Haw R, Loney F, May B, Milacic M, Rothfels K, Sevilla C, Shamovsky V, Shorser S, Varusai T, Weiser J, Wu G, Stein L, Hermjakob H, D'Eustachio P. The reactome pathway knowledgebase. Nucleic Acids Res. 2020 Jan 8;48(D1):D498-D503. doi: 10.1093/nar/gkz1031. PubMed PMID: 31691815.

Reactome Content Services API Documentation

Examples

# \donttest{
rba_reactome_query(ids = c("8953958", "11982506", "R-ALL-9649879"))
# }
# \donttest{
rba_reactome_query(ids = "R-HSA-9656256", enhanced = TRUE)
# }
# \donttest{
rba_reactome_query(ids = "8863054", attribute_name = "displayName")
# }