Map Cross References IDs to Reactome ReferenceEntity

Use this function To retrieve a list of Reactome ReferenceEntity associated to your provided Cross Reference (i.e. External) ID.

rba_reactome_xref(xref_id, ...)

Arguments

xref_id
molecule's cross-reference (external) identifier.

...
rbioapi option(s). Refer to rba_options's arguments documentation for more information on available options.

Value

List containing the ReferenceEntity corresponding to your provided cross-reference (external) ID.

Details

Reactome cross-references external database's identifiers to it's database Entries named ReferenceEntity, which resembles the invariant aspect of a molecule. Thus there is a one-to-many relationship between Reactome's ReferenceEntity object and the molecule's ID in external databases, which in Reactome's terms is called Cross Reference.
Refer to rba_reactome_participants's "Details section" to learn more about how Reactome classifies molecules.

Corresponding API Resources

"GET https://reactome.org/ContentService/references/mapping/identifier"

References

Jassal B, Matthews L, Viteri G, Gong C, Lorente P, Fabregat A, Sidiropoulos K, Cook J, Gillespie M, Haw R, Loney F, May B, Milacic M, Rothfels K, Sevilla C, Shamovsky V, Shorser S, Varusai T, Weiser J, Wu G, Stein L, Hermjakob H, D'Eustachio P. The reactome pathway knowledgebase. Nucleic Acids Res. 2020 Jan 8;48(D1):D498-D503. doi: 10.1093/nar/gkz1031. PubMed PMID: 31691815.

Reactome Content Services API Documentation

Examples

# \donttest{
rba_reactome_xref("CD40")
# }
# \donttest{
rba_reactome_xref("ENSP00000361350")
# }