Get Similarity Scores Hits of Proteins in Different Species

Using this function, you can retrieve highest Smith-Waterman bit scores among your input proteins and proteins in every other STRING species (e.g. the closest homologous protein of your input protein in other species). Bit Scores serve as similarity scores between protein sequence; And, according to STRING documentations, as a proxy for protein homology.

Usage

rba_string_homology_inter(ids, species = NULL, species_b = NULL, ...)

Arguments

ids

Your protein ID(s). It is strongly recommended to supply STRING IDs. See rba_string_map_ids for more information.

species

Numeric: NCBI Taxonomy identifier of your input proteins; Human Taxonomy ID is 9606. (Recommended, but optional if your input is less than 100 IDs.)

species_b

(optional) Numeric: one or more NCBI Taxonomy identifiers of species to limit the closets homologous proteins search.

...

rbioapi option(s). See rba_options's arguments manual for more information on available options.

Value

A data frame with Your input proteins and it's closest homologous proteins among all other (or a defined) STRING species.

Details

Note that this function will return the highest similarity score hits of your given protein(s) and their closets homologous proteins in other species. to retrieve similarity scores of different proteins within the same species see rba_string_homology_intra.
Similarity matrix is imported -by STRING- from: Similarity Matrix of Proteins (SIMAP)

Corresponding API Resources

"POST https://string-db.org/api/[output-format]/homology_best? identifiers=[your_identifiers]"

References

Szklarczyk D, Gable AL, Lyon D, Junge A, Wyder S, Huerta-Cepas J, Simonovic M, Doncheva NT, Morris JH, Bork P, Jensen LJ, Mering CV. STRING v11: protein-protein association networks with increased coverage, supporting functional discovery in genome-wide experimental datasets. Nucleic Acids Res. 2019 Jan 8;47(D1):D607-D613. doi: 10.1093/nar/gky1131. PMID: 30476243; PMCID: PMC6323986.

STRING API Documentation

See also

Examples

# \donttest{
rba_string_homology_inter(ids = "p53",
    species = 9606,
    species_b = 7070)
# }
# \donttest{
rba_string_homology_inter(ids = "ENSP00000269305", species = 9606)
# }