Get STRING Network Interactions

This function will retrieve Sting interaction pairs among your input protein ids, with the combined score and separate score for each STRING score channels. You can further expand your network to a defined size by providing "add_node" parameter.

rba_string_interactions_network(
  ids,
  species = NA,
  required_score = NA,
  add_nodes = NA,
  network_type = "functional",
  ...
)

Arguments

ids
Your protein IDs. It is strongly recommended to provide STRING IDs. See rba_string_map_ids for more information.

species
Numeric: NCBI Taxonomy identifier; Human Taxonomy ID is 9606. (Recommended, but optional if your input is less than 100 IDs.)

required_score
Numeric: A minimum of interaction score for an interaction to be included in the image. if not provided, the threshold will be applied by STRING Based in the network. (low Confidence = 150, Medium Confidence = 400, High Confidence = 700, Highest confidence = 900)

add_nodes
Numeric: Number of neighboring proteins to be added to the network. If none provided by the user, this argument value will depend on the number of provided "ids" argument:

Single id: add_node will be set to 10 to retrieve the interaction neighborhood of you input protein.

Multiple ids: add_node will be set to 0, thus the output will be the interactions between your input proteins.

network_type
should be one of:

"functional": (default) The edge's indicate both physical and functional associations.

"physical": The edges indicate that two proteins have a physical interaction or are parts of a complex.

...
rbioapi option(s). Refer to rba_options's arguments documentation for more information on available options.

Value

A data frame which each row is a network interaction and the columns contains interactor information and interaction scores:

stringId_A: STRING identifier (protein A)

stringId_B:STRING identifier (protein B)

preferredName_A: common protein name (protein A)

preferredName_B: common protein name (protein B)

ncbiTaxonId: NCBI taxon identifier

score: combined score

nscore: gene neighborhood score

fscore: gene fusion score

pscore: phylogenetic profile score

ascore: co-expression score

escore: experimental score

dscore: database score

tscore: textmining score

Details

Note that this function will return interactions between your set of provided proteins, or at most, expand the interaction network by the given parameters. TO retrieve a list of all possible interacting proteins with your given input, refer to rba_string_interaction_partners.

Corresponding API Resources

"POST https://string-db.org/api/[output-format]/network?identifiers= [your_identifiers]&[optional_parameters]"

References

Szklarczyk D, Gable AL, Lyon D, Junge A, Wyder S, Huerta-Cepas J, Simonovic M, Doncheva NT, Morris JH, Bork P, Jensen LJ, Mering CV. STRING v11: protein-protein association networks with increased coverage, supporting functional discovery in genome-wide experimental datasets. Nucleic Acids Res. 2019 Jan 8;47(D1):D607-D613. doi: 10.1093/nar/gky1131. PMID: 30476243; PMCID: PMC6323986.

STRING API Documentation

See also

Examples

# \donttest{
rba_string_interactions_network(ids = c("9606.ENSP00000269305",
    "9606.ENSP00000398698",
    "9606.ENSP00000275493"),
    network_type = "functional")
# }
# \donttest{
rba_string_interactions_network(ids = c("9606.ENSP00000269305",
    "9606.ENSP00000398698",
    "9606.ENSP00000275493"),
    species = 9606,
    add_nodes = 10)
# }