Get STRING Network Image

Depending on that you provided a single protein ID or more than one protein ID, this function will produce a static image of the interaction networks among your input proteins or/and with other proteins. Refer to the "Arguments" section to learn more about how you can modify the network image.

rba_string_network_image(
  ids,
  image_format = "image",
  save_image = TRUE,
  species = NA,
  add_color_nodes = NA,
  add_white_nodes = NA,
  required_score = NA,
  network_flavor = "confidence",
  network_type = "functional",
  hide_node_labels = FALSE,
  hide_disconnected_nodes = FALSE,
  hide_structure_pics = FALSE,
  ...
)

Arguments

ids
Your protein ID(s). It is strongly recommended to provide STRING IDs. See rba_string_map_ids for more information.

image_format
one of:

"image": PNG image with normal resolution.

"highres_image": High-resolution PNG image.

"svg": Scalable Vector Graphics image.

save_image
Logical or Character:

TRUE: Save the image to an automatically-generated path.

FALSE: Do not save the image, just return it as an R object.

Character string: A valid file path to save the image to.

species
Numeric: NCBI Taxonomy identifier; Human Taxonomy ID is 9606. (Recommended, but optional if your input is less than 100 IDs.)

add_color_nodes
Numeric: The number of colored nodes (queried proteins and first shell of interactors) to be added.

add_white_nodes
Numeric: The number of white nodes (second shell of interactors) to be added after colored nodes.

required_score
Numeric: A minimum of interaction score for an interaction to be included in the image. if not provided, the threshold will be applied by STRING Based in the network. (low Confidence = 150, Medium Confidence = 400, High Confidence = 700, Highest confidence = 900)

network_flavor
The style of network edges, should be one of:

"confidence": (default) Line's thickness is an indicator of the interaction's confidence score.

"evidence": Line's color is based on the type of evidences that support the interaction.

"action": Line's Shape is an indicator of the interaction's predicted mode of actions.

network_type
should be one of:

"functional": (default) The edge's indicate both physical and functional associations.

"physical": The edges indicate that two proteins have a physical interaction or are parts of a complex.

hide_node_labels
Logical: (Default = FALSE) Hide proteins names from the image?

hide_disconnected_nodes
Logical: (Default = FALSE) Hide proteins that are not connected to any other proteins from the image?

hide_structure_pics
Logical: (Default = FALSE) Hide protein's structure picture from inside the bubbles?

...
rbioapi option(s). Refer to rba_options's arguments documentation for more information on available options.

Value

A network images which can be PNG or SVG depending on the inputs.

Corresponding API Resources

"POST https://string-db.org/api/[output-format]/network?identifiers= [your_identifiers]&[optional_parameters]"

References

Szklarczyk D, Gable AL, Lyon D, Junge A, Wyder S, Huerta-Cepas J, Simonovic M, Doncheva NT, Morris JH, Bork P, Jensen LJ, Mering CV. STRING v11: protein-protein association networks with increased coverage, supporting functional discovery in genome-wide experimental datasets. Nucleic Acids Res. 2019 Jan 8;47(D1):D607-D613. doi: 10.1093/nar/gky1131. PMID: 30476243; PMCID: PMC6323986.

STRING API Documentation

See also

Examples

if (FALSE) {
rba_string_network_image(ids = c("9606.ENSP00000269305",
    "9606.ENSP00000398698",
    "9606.ENSP00000275493"),
    network_type = "functional",
    save_image = FALSE)
}
if (FALSE) {
rba_string_network_image(ids = c("TP53", "TNF", "EGFR"),
    species = 9606,
    save_image = TRUE)
}
if (FALSE) {
rba_string_network_image(ids = "9606.ENSP00000269305",
    image_format = "highres_image",
    save_image = file.path(getwd(), "TP53_network.png"))
}