Search Proteomics Peptides in UniProt

UniProt maps proteomics peptides from different sources to the proteins' sequences. Using this function, you can search for proteomics peptides that has been map to UniProt proteins. You may also refine your search with modifiers such as data_source, peptide etc. See "Arguments section" for more information.

Usage

rba_uniprot_proteomics_search(
  accession = NULL,
  data_source = NULL,
  taxid = NULL,
  upid = NULL,
  peptide = NULL,
  unique = NULL,
  ...
)

Arguments

accession

UniProtKB primary or secondary accession(s). You can supply up to 100 accession numbers.

data_source

Proteomics data source. You can choose up to two of:

"MaxQB"

"PeptideAtlas"

"EPD"

"ProteomicsDB"

taxid

NIH-NCBI Taxon ID. You can supply up to 20 taxon IDs.

upid

UniProt Proteome identifier (UPID). You can supply up to 100 UPIDs.

peptide

Peptide sequence(s). You can supply up to 20 sequences.

unique

Logical: Should the results be filtered based on the Peptide's uniqueness (the fact that a peptide maps to only 1 protein). If TRUE, Only unique peptides will be returned, if FALSE only un-unique peptides will be returned; If NULL (default) the results will not be filtered based on this.

...

rbioapi option(s). See rba_options's arguments manual for more information on available options.

Value

A list Where each element correspond to a UniProt protein and proteomics peptides are organized under the "features" sub-list.

Details

Note that this is a search function. Thus, you are not required to fill every argument; You may use whatever combinations of arguments you see fit for your query.
see also: Mass spectrometry-based proteomics data in UniProtKB

Corresponding API Resources

"GET https://www.ebi.ac.uk/proteins/api/proteomics"

References

The UniProt Consortium, UniProt: the universal protein knowledgebase in 2021, Nucleic Acids Research, Volume 49, Issue D1, 8 January 2021, Pages D480–D489, https://doi.org/10.1093/nar/gkaa1100

Andrew Nightingale, Ricardo Antunes, Emanuele Alpi, Borisas Bursteinas, Leonardo Gonzales, Wudong Liu, Jie Luo, Guoying Qi, Edd Turner, Maria Martin, The Proteins API: accessing key integrated protein and genome information, Nucleic Acids Research, Volume 45, Issue W1, 3 July 2017, Pages W539–W544, https://doi.org/10.1093/nar/gkx237

Proteins API Documentation

Citations note on UniProt website

See also

Examples

# \donttest{
rba_uniprot_proteomics_search(peptide = "MEDYTKIEK")
# }
# \donttest{
rba_uniprot_proteomics_search(peptide = "MEDYTKIEK")
# }
if (FALSE) { # \dontrun{
### this will generate a very large response!
  rba_uniprot_proteomics_search(taxid = 9606,
  data_source = "PeptideAtlas",
  progress = TRUE, timeout = 999999, unique = TRUE)
} # }