Search UniParc Entries

Use this function to search UniProt Archive (UniParc) entries.You may also refine your search with modifiers such as sequence length, taxon id etc. refer to "Arguments section" for more information.

rba_uniprot_uniparc_search(
  upi = NA,
  accession = NA,
  db_type = NA,
  db_id = NA,
  gene = NA,
  protein = NA,
  taxid = NA,
  organism = NA,
  sequence_checksum = NA,
  ipr = NA,
  signature_db = NA,
  signature_id = NA,
  upid = NA,
  seq_length = NA,
  rf_dd_type = NA,
  rf_db_id = NA,
  rf_active = NA,
  rf_tax_id = NA,
  ...
)

Arguments

upi
unique UniParc Identifier(s). You can provide up to 100 IDs.

accession
UniProtKB primary or secondary accession(s). You can provide up to 100 accession numbers.

db_type
cross-reference (external database) name.

db_id
Protein ID in the cross-reference (external) database. You can provide up to 100 IDs.

gene
UniProt gene name(s). You can provide up to 20 gene names.

protein
UniProt protein name.

taxid
NIH-NCBI Taxon ID. You can provide up to 20 taxon IDs.

organism
Organism name.

sequence_checksum
Sequence CRC64 checksum.

ipr
InterPro identifier(s). You can provide up to 20 IDs.

signature_db
InterPro's signature database. You can provide up to 13 of the following values:
"CATH", "CDD", "HAMAP", "MobiDB Lite", "Panther", "Pfam", "PIRSF", "PRINTS", "Prosite", "SFLD", "SMART", "SUPERFAMILY" and/or "TIGRfams"

signature_id
Signature ID in the InterPro's signature database. You can provide up to 20 IDs.

upid
UniProt Proteome identifier (UPID). You can provide up to 100 UPIDs.

seq_length
An exact sequence length (e.g. 150) or a range of sequence lengths (e.g. "130-158").

rf_dd_type
Filter the content of the each UniParc entry by cross-reference names. You can provide multiple values.

rf_db_id
Filter the content of the each UniParc entry by protein identifiers in any cross-reference database. You can provide multiple values.

rf_active
(logical ) Filter the content of each UniParc entry based on active status on source database:

NA: don't filter contents based on active status.

TRUE: only return contents which are still active.

FALSE: Only return contents which are not active.

rf_tax_id
(Numeric) Filter the content of each UniParc entry by NIH-NCBI Taxon ID. You can provide multiple values.

...
rbioapi option(s). Refer to rba_options's arguments documentation for more information on available options.

Value

A List where each element corresponds to one UniParc entry returned by your search query. The element itself is a sub-list containing sequence information and reference entries.

Details

Note that this is a search function. Thus, you are not required to fill every argument; You may use whatever combinations of arguments you see fit for your query.

Corresponding API Resources

"GET https://ebi.ac.uk/proteins/api/uniparc"

References

Andrew Nightingale, Ricardo Antunes, Emanuele Alpi, Borisas Bursteinas, Leonardo Gonzales, Wudong Liu, Jie Luo, Guoying Qi, Edd Turner, Maria Martin, The Proteins API: accessing key integrated protein and genome information, Nucleic Acids Research, Volume 45, Issue W1, 3 July 2017, Pages W539–W544, https://doi.org/10.1093/nar/gkx237

Proteins API Documentation

See also

Examples

# \donttest{
rba_uniprot_uniparc_search(upi = "UPI00000000C9")
# }
# \donttest{
rba_uniprot_uniparc_search(accession = "P30914")
# }
# \donttest{
rba_uniprot_uniparc_search(accession = "P30914", rf_active = TRUE)
# }
# \donttest{
rba_uniprot_uniparc_search(taxid = "694009", protein = "Nucleoprotein")
# }